+++ NackEngine/core/physics/HitStruct.cs
@@ -11,18 +11,17 @@ namespace NackEngine.core.physics
     {
         public Point Point { get; set; }
         public NVector Normal { get; set; }
-        public double T { get; set; }
+        public float T { get; set; }
 
-        // Cordinates 2D (U, V)
-        public double U { get; set; }
+        public float U { get; set; }
 
-        public double V { get; set; }
+        public float V { get; set; }
 
         public bool FrontFace { get; set; }
         public Material Material { get; set; }
 
         public void setFaceNormal(Ray ray, NVector owNormal) {
-            this.FrontFace = NVector.Dot(ray.Direction(), owNormal) <0;
+            this.FrontFace = NVector.Dot(ray.Direction(), owNormal) < 0;
             this.Normal = FrontFace? owNormal : -owNormal;
         }
     }
