+++ NackEngine/core/render/Material.cs
@@ -10,13 +10,13 @@ namespace NackEngine.core.render
     {
         public bool Bounce(Ray ray, HitStruct hit, out ScatterStruct scatter);
 
-        public double ScatterProb(Ray ray, HitStruct hit, Ray scattered) {
-            return 0;
+        public float ScatterProb(Ray ray, HitStruct hit, Ray scattered) {
+            return 0f;
         }
 
-        public Color Emitted(double u, double v, Point point)
+        public Color Emitted(float u, float v, Point point)
         {
-            return new Color(0, 0, 0);
+            return new Color(0f, 0f, 0f);
         }
     }
 }
