+++ NackEngine/math/ProbDensity.cs
@@ -7,7 +7,7 @@ namespace NackEngine.math
 {
     public interface ProbDensity
     {
-        double Value(NVector direction);
+        float Value(NVector direction);
         NVector Generate();
     }
 }
